Changes signal and store modules into just one (reactive)
This change was done since both of the modules are very related and maybe should be in a single module but neither signal nor store fit both so a change was done and the reactive module was created. This also document with jsDoc very well all the public, private and internal code.